requirements.txt
einops
exifread
numpy==1.24.1
rawpy==0.20.0
tensorboard
timm
thop
torchmetrics

